Implement usePlayerIntentHandler
